% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{as.promise.mirai}
\alias{as.promise.mirai}
\title{Make mirai promise}
\usage{
\method{as.promise}{mirai}(x)
}
\arguments{
\item{x}{an object of class 'mirai'.}
}
\value{
A 'promise' object.
}
\description{
Creates a 'promise' from a 'mirai'.
}
\details{
This function is an S3 method for the generic \code{\link{as.promise}}
    for class 'mirai'.
}
\examples{
if (interactive()) {
# Only run examples in interactive R sessions

p <- promises::as.promise(mirai::mirai("example"))
print(p)
promises::is.promise(p)

}

}
