% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{polling}
\alias{polling}
\title{Set Polling Frequency}
\usage{
polling(freq = 100L)
}
\arguments{
\item{freq}{[default 100L] integer number of milliseconds.}
}
\value{
Invisible NULL.
}
\description{
Set the frequency with which to poll for promise resolution (the default
    being 100 milliseconds).
}
\examples{
# set polling frequency to 1s
polling(freq = 1000L)

# reset polling frequency to default 100 ms
polling()

}
