% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{as.promise.mirai}
\alias{as.promise.mirai}
\alias{as.promise.recvAio}
\title{Make 'Mirai' 'Promise'}
\usage{
\method{as.promise}{mirai}(x)

\method{as.promise}{recvAio}(x)
}
\arguments{
\item{x}{an object of class 'mirai' or 'recvAio'.}
}
\value{
A 'promise' object.
}
\description{
Creates a 'promise' from a 'mirai' or 'recvAio'.
}
\details{
This function is an S3 method for the generic \code{\link{as.promise}}
    for class 'mirai' or 'recvAio'.

    Set the option \code{mirai.promises} at any time using
    \code{options(mirai.promises = <numeric value>)} to provide a frequency
    (in seconds) other than the default 0.1s with which to poll for promise
    resolution.
}
\examples{
if (interactive()) {
# Only run examples in interactive R sessions

p <- promises::as.promise(mirai::mirai("example"))
print(p)
promises::is.promise(p)

}

}
