% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{as.promise.mirai}
\alias{as.promise.mirai}
\alias{as.promise.recvAio}
\title{Make 'Mirai' 'Promise'}
\usage{
\method{as.promise}{mirai}(x)

\method{as.promise}{recvAio}(x)
}
\arguments{
\item{x}{an object of class 'mirai' or 'recvAio'.}
}
\value{
A 'promise' object.
}
\description{
Creates a 'promise' from a 'mirai' or 'recvAio'.
}
\details{
This function is an S3 method for the generic \code{\link{as.promise}}
    for class 'mirai' or 'recvAio'.

    \code{\link{polling}} may be used to customise the frequency with which
    to poll for promise resolution (defaults to every 100 milliseconds).
}
\examples{
if (interactive()) {
# Only run examples in interactive R sessions

p <- promises::as.promise(mirai::mirai("example"))
print(p)
promises::is.promise(p)

}

}
