% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirai.promises-package.R
\docType{package}
\encoding{UTF-8}
\name{mirai.promises-package}
\alias{mirai.promises-package}
\title{mirai.promises: Make 'Mirai' 'Promises'}
\description{
Use 'mirai' as 'promises' for easy integration in 'plumber' or 'shiny'
    pipelines.
}
\section{Links}{


    mirai website: \url{https://shikokuchuo.net/mirai/}\cr
    mirai on CRAN: \url{https://cran.r-project.org/package=mirai}
}

\author{
Charlie Gao \email{charlie.gao@shikokuchuo.net}
    (\href{https://orcid.org/0000-0002-0750-061X}{ORCID})
}
